mapping tables

+++ sourceCode/backend/greenbasketdemo/src/main/java/com/app/dao/ProductRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.app.pojos.ProductDetails;
 
 
-public interface ProductRepository extends JpaRepository<ProductDetails, Integer> {
+public interface ProductRepository extends JpaRepository<ProductDetails, Long> {
